requirements.txt
Django==2.1.7
gunicorn==19.9.0
django-heroku==0.3.1
psycopg2==2.7.6.1
django-webpack-loader==0.6.0
django-nose==1.4.6
nose==1.3.7
factory_boy==2.11.1
django-debug-toolbar==1.11
git+git://github.com/romack77/django-cprofile-middleware.git@65d85a7e35c582c811ce501fe20ae54733674f2e
